requirements.txt
flask
flask-restx
scrapy
selenium
pytest-flask
python-dotenv